% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniparquet-package.R
\docType{package}
\name{miniparquet-package}
\alias{miniparquet}
\alias{miniparquet-package}
\title{miniparquet: Read Parquet Files}
\description{
Self-sufficient reader for a subset of Parquet files. Nested tables, compression besides Snappy and encryption are not supported.
}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Hannes Mühleisen (\href{https://orcid.org/0000-0001-8552-0029}{ORCID})
}

Other contributors:
\itemize{
  \item Google Inc. [copyright holder]
  \item Apache Software Foundation [copyright holder]
  \item Daniel Lemire [copyright holder]
  \item Chad Walters [copyright holder]
  \item Posit Software, PBC [copyright holder]
}

}
\keyword{internal}
