% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{parquet_schema}
\alias{parquet_schema}
\title{Create a Parquet schema}
\usage{
parquet_schema(...)
}
\arguments{
\item{...}{Parquet type specifications, see below.
For backwards compatibility, you can supply a file name
here, and then \code{parquet_schema} behaves as \code{\link[=read_parquet_schema]{read_parquet_schema()}}.}
}
\description{
Create a Parquet schema
}
