% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniparquet.R
\name{parquet_write}
\alias{parquet_write}
\alias{write_parquet}
\title{Write a data frame to a Parquet file}
\usage{
parquet_write(x, file, compression = c("snappy", "uncompressed"))

write_parquet(x, file, compression = c("snappy", "uncompressed"))
}
\arguments{
\item{x}{Data frame to write.}

\item{file}{Path to the output file.}

\item{compression}{Compression algorithm to use. Currently only
\code{"snappy"} (the default) and \code{"uncompressed"} are supported.}
}
\value{
\code{NULL}
}
\description{
Writes the contents of an R data frame into a Parquet file.
}
\examples{
\dontshow{if (!miniparquet:::is_rcmd_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# add row names as a column, because `parquet_write()` ignores them.
mtcars2 <- cbind(name = rownames(mtcars), mtcars)
parquet_write(mtcars2, "mtcars.parquet")
\dontshow{\}) # examplesIf}
}
