% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parquet2}
\alias{read_parquet2}
\title{Read Parquet file, next gen}
\usage{
read_parquet2(file, options = parquet_options())
}
\arguments{
\item{file}{Path to a Parquet file.}

\item{options}{Nanoparquet options, see \code{\link[=parquet_options]{parquet_options()}}.}
}
\description{
Read Parquet file, next gen
}
