% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porcelain.R
\name{rle_encode_int}
\alias{rle_encode_int}
\title{RLE encode integers}
\usage{
rle_encode_int(x)
}
\arguments{
\item{x}{Integer vector.}
}
\value{
Raw vector, the encoded integers. It has two attributes:
\itemize{
\item \code{bit_length}: the number of bits needed to encode the input, and
\item \code{length}: length of the original integer input.
}
}
\description{
RLE encode integers
}
\seealso{
\code{\link[=rle_decode_int]{rle_decode_int()}}

Other encodings: 
\code{\link{rle_decode_int}()}
}
\concept{encodings}
\keyword{internal}
