% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniparquet.R
\name{write_parquet}
\alias{write_parquet}
\title{Write a data frame to a Parquet file}
\usage{
write_parquet(x, file, compression = c("snappy", "uncompressed"))
}
\arguments{
\item{x}{Data frame to write.}

\item{file}{Path to the output file.}

\item{compression}{Compression algorithm to use. Currently only
\code{"snappy"} (the default) and \code{"uncompressed"} are supported.}
}
\value{
\code{NULL}
}
\description{
Writes the contents of an R data frame into a Parquet file.
}
\examples{
\dontshow{if (!miniparquet:::is_rcmd_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# add row names as a column, because `write_parquet()` ignores them.
mtcars2 <- cbind(name = rownames(mtcars), mtcars)
write_parquet(mtcars2, "mtcars.parquet")
\dontshow{if (Sys.getenv("NOT_CRAN") == "true") unlink("mtcars.parquet")}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=parquet_metadata]{parquet_metadata()}}, \code{\link[=read_parquet]{read_parquet()}}.
}
