% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniparquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a Parquet file into a data frame}
\usage{
read_parquet(file)
}
\arguments{
\item{file}{Path to a Parquet file.}
}
\value{
A \code{data.frame} with the file's contents.
}
\description{
Converts the contents of the named Parquet file to a R data frame.
}
\examples{
file_name <- system.file("extdata/userdata1.parquet", package = "miniparquet")
parquet_df <- miniparquet::read_parquet(file_name)
print(str(parquet_df))
}
\seealso{
\code{\link[=read_parquet_metadata]{read_parquet_metadata()}}, \code{\link[=write_parquet]{write_parquet()}}.
}
